+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Version History
 
+## 0.1.0.dev0
+
+* Developer release (not public to pip)
+* Improve tests for CSVPack
+* Update README.md
+* Change pyo3 revision to 0.28.2
+* Refactor deprecated Rust code
+* Refactor CSVPackMeta class
+* Improve csvpack_repr() function
+* Refactor CSVPackReader.`__repr__`() method
+* Refactor CSVPackWriter.`__repr__`() method
+* Update depends light-compressor==0.1.1.dev0
+
 ## 0.0.0.1
 
 First version of the library (No public to pip)